requirements.txt
notion==0.0.28
django==4.1.3
python-telegram-bot==13.14
